Added Railfence & Column Transposition Cipher
The rail fence cipher is an easy to apply transposition cipher that jumbles up the order of the letters of a message in a quick convenient way.
The Column Transposition Cipher is a form of transposition cipher just like Rail Fence Cipher. Columnar Transposition involves writing the plaintext out in rows, and then reading the ciphertext off in columns one by one.